Upload updated/bug-fixed version of music-player.sh
This update consists of many bux fixes as well as a rewrite of the configur
tion script. Now, the script does not overstep its boundaries by deleting an
d moving files -- it simply changes the appropriate lines in 'teapot'. By de
fault, teapot uses ffmpeg, but if the user is on a Mac the configuration scr
ipt will change that to afplay. I've also changed some key mappings in teapo
t, and added a PID global variable containing the process id of the song pla
ying. Bugs regarding EXTENSIONS were fixed -- ext_list is no longer in exist
ence, and EXTENSIONS is an array (not relying on any dodgy behavior). The me
chanism for num_songs is different (and actually functional). show_music_pla
yer's nonsensical code was replaced (now it simply relays messages to show_o
ptions, which can either 'reset', set seconds to a value, or just show optio
ns with the default amount of seconds. Pausing works but is still dependent
on the process not being killed randomly by the system (which means you can't pause for extensive amounts of time). That's about it!